+++ extensions/vertx-http/deployment/src/main/java/io/quarkus/devui/deployment/welcome/WelcomeProcessor.java
@@ -72,11 +72,11 @@ public class WelcomeProcessor {
                     Path resourceDirs = resourcesDirs.iterator().next().getDir();
                     Path propertiesFile = resourceDirs.resolve("application.properties");
                     if (Files.exists(propertiesFile)) {
-                        return propertiesFile.toString().replace("\\", "/").substring((int) root.length() + 1);
+                        return propertiesFile.toString().replace("\\", "/").substring(root.length() + 1);
                     }
                     Path propertiesYaml = resourceDirs.resolve("application.yaml");
                     if (Files.exists(propertiesYaml)) {
-                        return propertiesYaml.toString().replace("\\", "/").substring((int) root.length() + 1);
+                        return propertiesYaml.toString().replace("\\", "/").substring(root.length() + 1);
                     }
                 }
             }
@@ -92,7 +92,7 @@ public class WelcomeProcessor {
                 Collection<SourceDir> sourceDirs = workspaceModule.getMainSources().getSourceDirs();
                 if (sourceDirs != null && !sourceDirs.isEmpty()) {
                     String sourceDir = sourceDirs.iterator().next().getDir().toString().replace("\\", "/");
-                    return sourceDir.substring((int) root.length() + 1);
+                    return sourceDir.substring(root.length() + 1);
                 }
             }
         }
@@ -107,7 +107,7 @@ public class WelcomeProcessor {
                 Collection<SourceDir> resourcesDirs = workspaceModule.getMainSources().getResourceDirs();
                 if (resourcesDirs != null && !resourcesDirs.isEmpty()) {
                     String resourceDirs = resourcesDirs.iterator().next().getDir().toString().replace("\\", "/");
-                    return resourceDirs.substring((int) root.length() + 1);
+                    return resourceDirs.substring(root.length() + 1);
                 }
             }
         }
