+++ extensions/websockets-next/deployment/src/test/java/io/quarkus/websockets/next/test/closereason/ClientCloseReasonTest.java
@@ -46,7 +46,7 @@ public class ClientCloseReasonTest {
                 .create()
                 .baseUri(closingUri)
                 .onClose((c, cr) -> {
-                    closeStatusCode.set((int) cr.getCode());
+                    closeStatusCode.set(cr.getCode());
                     closeMessage.set(cr.getMessage());
                     closedClientLatch.countDown();
                 })
